Rearranging the files.
