+++ Honehoover/Honehoover/highscores.txt
@@ -1,5 +1,5 @@
+12 Easy
 14 Easy
 113 Easy
 213 Easy
 313 Easy
-413 Easy
